Split up utils functions into more logical modules
