+++ src/streetscapes/__init__.py
@@ -1,20 +1,6 @@
 """Streetscapes python package for large-scale analysis of street-level imagery."""
 
-import os
-
-from dotenv import load_dotenv
-
 from streetscapes.config import CFG
-from streetscapes.utils import logger
-
-
-def get_env(key: str):
-    """Get environment variable from user's env."""
-    load_dotenv()
-    value = os.getenv(key, None)
-    if value is not None:
-        return value
-    raise KeyError(f"{key} not found in environment variables.")
-
+from streetscapes.utils import get_env, logger
 
 __all__ = ["CFG", "logger", "get_env"]
