+++ src/streetscapes/utils/__init__.py
@@ -1,44 +1,35 @@
 """Utilities."""
 
-from streetscapes.utils.functions import (
-    as_hsv,
-    as_rgb,
-    camel2snake,
+from streetscapes.utils.env import get_env
+from streetscapes.utils.exif import extract_exif_data
+from streetscapes.utils.geo import get_geohash_shard_path
+from streetscapes.utils.images import as_hsv, as_rgb, open_image
+from streetscapes.utils.logging import logger
+from streetscapes.utils.metadata import get_image_metadata
+from streetscapes.utils.paths import (
     ensure_dir,
-    extract_categories,
-    extract_exif_data,
     filter_files,
-    get_device,
-    get_env,
-    get_geohash_shard_path,
-    get_image_hash,
-    get_image_metadata,
     get_image_paths,
+    hide_home,
+    make_path,
+    show_dir_tree,
+)
+from streetscapes.utils.plotting import make_colourmap, plot_metadata
+from streetscapes.utils.time import iso_timestamp
+from streetscapes.utils.uuids import (
+    get_image_hash,
     get_image_uuid,
     hash2uuid,
-    hide_home,
-    is_notebook,
-    iso_timestamp,
-    make_colourmap,
-    make_path,
-    open_image,
-    plot_metadata,
-    show_dir_tree,
-    show_image,
     uuid7,
 )
-from streetscapes.utils.logging import logger
 
 __all__ = [
     "logger",
     "as_hsv",
     "as_rgb",
-    "camel2snake",
     "ensure_dir",
-    "extract_categories",
     "extract_exif_data",
     "filter_files",
-    "get_device",
     "get_env",
     "get_geohash_shard_path",
     "get_image_hash",
@@ -47,13 +38,11 @@ __all__ = [
     "get_image_uuid",
     "hash2uuid",
     "hide_home",
-    "is_notebook",
     "iso_timestamp",
     "make_colourmap",
     "make_path",
     "open_image",
     "plot_metadata",
     "show_dir_tree",
-    "show_image",
     "uuid7",
 ]
